Juggled <script> tags
Moved <script> tags around for a while, until we figured out the
correct place. The problem is separating the base.html code from the
one specific to the homepage, because we started with one initial
homepage that we made and then converted that to a website base file.

These things happen.