Basic AppDetail page created with dummy screenshots
For it to be responsive you'll have to reload the page in different
dimentions because the javascript checks the dimentions only when
the page loads.

The Download button also needs to be made functional as well as the
download size. The screenshots also need to be added to the slider.
Also the Idea section needs to be filled with the ideabase.